apps/queue-trigger/pyproject.toml
[tool.poetry]
name = "queue-trigger"
version = "0.1.0"
description = ""
authors = ["Masaya Sonobe <masayasonobe@microsoft.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
azure-functions = "^1.20.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
